# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nestora API - Business platform for small businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List support tickets (owner sees all; customer sees own)
      responses:
        "200":
          description: Ticket list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a support ticket (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/reply:
    post:
      operationId: replyToTicket
      tags: [tickets]
      summary: Reply to a support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyInput"
      responses:
        "201":
          description: Reply added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketReply"

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TicketReply:
      type: object
      required: [id, ticketId, authorId, authorName, authorRole, message, createdAt]
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorRole:
          type: string
          enum: [owner, customer]
        message:
          type: string
        createdAt:
          type: string
    Ticket:
      type: object
      required: [id, customerId, customerName, subject, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        replies:
          type: array
          items:
            $ref: "#/components/schemas/TicketReply"
        createdAt:
          type: string
        updatedAt:
          type: string
    TicketInput:
      type: object
      required: [subject, message]
      properties:
        subject:
          type: string
          minLength: 3
        message:
          type: string
          minLength: 10
    TicketReplyInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1

    # NOTIFICATIONS
...